requirements.txt
# ── Core ───────────────────────────────────────────
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic[email]>=2.0,<3.0
pydantic-settings>=2.0,<3.0

# ── HTTP client ────────────────────────────────────
httpx>=0.27.0,<1.0.0

# ── Google APIs & OAuth ────────────────────────────
google-api-python-client>=2.130.0,<3.0.0
google-auth>=2.30.0,<3.0.0
google-auth-oauthlib>=1.2.0,<2.0.0
google-auth-httplib2>=0.2.0,<1.0.0

# ── JWT ────────────────────────────────────────────
PyJWT>=2.8.0,<3.0.0

# ── Dev / QA (optional) ───────────────────────────
python-dotenv>=1.0.0
